src/requirements.txt
requests==2.32.5
psutil==7.1.3
docker==7.1.0
pyyaml==6.0.3
rich==14.2.0